requirements.txt
streamlit>=1.31
biopython>=1.83
rdkit-pypi>=2022.9.5
pillow>=9.5
numpy>=1.23
